requirement.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
apipkg==1.5
astor==0.8.1
async-generator==1.10
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==3.4.7
et-xmlfile==1.1.0
execnet==1.8.1
h11==0.12.0
idna==2.10
iniconfig==1.1.1
openpyxl==3.0.7
outcome==1.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytest-ast-transformer==1.0.3
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-soft-assertions==0.1.2
pytest-xdist==2.2.1
requests==2.25.1
selenium==4.0.0b3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
texttable==1.6.3
toml==0.10.2
trio==0.18.0
trio-websocket==0.9.2
urllib3==1.26.5
webdriver-manager==3.4.2
wsproto==1.0.0
